multi core on wasm

+++ golem/task/taskserver.py
@@ -1,4 +1,6 @@
-# -*- coding: utf-8 -*-
+# pylint: disable=too-many-instance-attributes,too-many-public-methods,
+# pylint: disable=too-many-lines
+
 import asyncio
 import functools
 import itertools
@@ -7,7 +9,6 @@ import os
 import shutil
 import time
 import weakref
-from dataclasses import dataclass, field
 from enum import Enum
 from pathlib import Path
 from typing import (
@@ -20,6 +21,7 @@ from typing import (
     Tuple,
     TYPE_CHECKING,
 )
+from dataclasses import dataclass, field
 
 from golem_messages import exceptions as msg_exceptions
 from golem_messages import message
@@ -70,6 +72,7 @@ from golem.ranking.manager.database_manager import (
     update_requestor_assigned_sum,
     update_requestor_efficiency,
 )
+from golem.resource.resourcehandshake import ResourceHandshake
 from golem.resource.resourcemanager import ResourceManager
 from golem.rpc import utils as rpc_utils
 from golem.task import helpers as task_helpers
@@ -207,21 +210,21 @@ class TaskServer(
         self.max_trust = 1.0
         self.min_trust = 0.0
 
-        self.last_messages = []
+        self.last_messages: List[Any] = []
 
-        self.results_to_send = {}
-        self.failures_to_send = {}
+        self.results_to_send: Dict[str, Any] = {}
+        self.failures_to_send: Dict[str, Any] = {}
 
         self.use_ipv6 = use_ipv6
 
         self.forwarded_session_request_timeout = \
             config_desc.waiting_for_task_session_timeout
-        self.forwarded_session_requests = {}
+        self.forwarded_session_requests: Dict[str, Any] = {}
         self.acl = get_acl(
             self.client, max_times=config_desc.disallow_id_max_times)
         self.acl_ip = DenyAcl(
             self.client, max_times=config_desc.disallow_ip_max_times)
-        self.resource_handshakes = {}
+        self.resource_handshakes: Dict[str, ResourceHandshake] = {}
         self.requested_tasks: Set[str] = set()
         self._last_task_request_time: float = time.time()
 
@@ -334,12 +337,18 @@ class TaskServer(
                 < self.config_desc.task_request_interval:
             return
 
-        if self.task_computer.has_assigned_task() \
-                or (not self.task_computer.compute_tasks) \
+        if (not self.task_computer.compute_tasks) \
                 or (not self.task_computer.runnable):
             return
 
-        task_header = self.task_keeper.get_task(self.requested_tasks)
+        if not self.task_computer.can_take_work():
+            return
+
+        compatibile_tasks = self.task_computer.compatible_tasks(
+            set(self.task_keeper.supported_tasks))
+
+        task_header = self.task_keeper.get_task(
+            exclude=self.requested_tasks, supported_tasks=compatibile_tasks)
         if task_header is None:
             return
 
@@ -357,6 +366,7 @@ class TaskServer(
         deferred.addErrback(_request_task_error)  # pylint: disable=no-member
 
     @inlineCallbacks
+    # pylint: disable=too-many-return-statements,too-many-branches
     def _request_task(self, theader: dt_tasks.TaskHeader) -> Deferred:
         try:
             supported = self.should_accept_requestor(theader.task_owner.key)
@@ -398,11 +408,16 @@ class TaskServer(
                     )
                 return None
 
+            num_subtasks = 1
             # Check performance
             if isinstance(env, OldEnv):
                 benchmark_result = env.get_benchmark_result()
                 benchmark_score = benchmark_result.performance
                 benchmark_cpu_usage = benchmark_result.cpu_usage
+                if env.is_single_core():
+                    num_subtasks = self.task_computer.free_cores
+                    if num_subtasks == 0:
+                        return None
             else:  # NewEnv
                 try:
                     future = asyncio.run_coroutine_threadsafe(
@@ -452,10 +467,7 @@ class TaskServer(
                 ProviderPricing(
                     price_per_wallclock_h=self.config_desc.min_price,
                     price_per_cpu_h=self.config_desc.price_per_cpu_h,
-                ),
-                theader.max_price,
-                theader.task_owner.key
-            )
+                ), theader.max_price, theader.task_owner.key)
             self.task_manager.add_comp_task_request(
                 theader=theader, price=price,
                 performance=benchmark_score
@@ -483,7 +495,7 @@ class TaskServer(
             return theader.task_id
         except Exception as err:  # pylint: disable=broad-except
             logger.warning("Cannot send request for task: %s", err)
-            logger.debug("Detailed traceback", exc_info=True)
+            logger.warning("Detailed traceback", exc_info=True)
             self.remove_task_header(theader.task_id)
 
         return None
@@ -492,7 +504,7 @@ class TaskServer(
             self,
             msg: message.tasks.TaskToCompute,
     ) -> bool:
-        if self.task_computer.has_assigned_task():
+        if not self.task_computer.can_take_work():
             logger.error("Trying to assign a task, when it's already assigned")
             return False
 
@@ -518,7 +530,8 @@ class TaskServer(
 
             defer.gatherResults(deferred_list, consumeErrors=True)\
                 .addCallbacks(
-                    lambda _: self.resource_collected(msg.task_id),
+                    lambda _: self.resource_collected(msg.task_id,
+                                                      msg.subtask_id),
                     lambda e: self.resource_failure(msg.task_id, e))
         else:
             self.request_resource(
@@ -573,7 +586,7 @@ class TaskServer(
             raise RuntimeError("Incorrect subtask_id: {}".format(subtask_id))
 
         # this is purely for tests
-        if self.config_desc.overwrite_results:
+        if self.config_desc.overwrite_results and result is not None:
             for file_path in result:
                 shutil.copyfile(
                     src=self.config_desc.overwrite_results,
@@ -584,11 +597,15 @@ class TaskServer(
         delay_time = 0.0
         last_sending_trial = 0
         stats = stats or {}
+        if result is None:
+            task_result: Tuple = (str(task_api_result),)
+        else:
+            task_result = tuple(result)
 
         wtr = WaitingTaskResult(
             task_id=task_id,
             subtask_id=subtask_id,
-            result=result or (str(task_api_result),),
+            result=task_result,
             last_sending_trial=last_sending_trial,
             delay_time=delay_time,
             owner=header.task_owner,
@@ -620,8 +637,7 @@ class TaskServer(
             task_api_result,
             client_options)
         deferred.addCallbacks(  # pylint: disable=no-member
-            on_result_share_success,
-            on_result_share_error)
+            on_result_share_success, on_result_share_error)
 
     def _create_and_set_result_package(self, wtr):
         task_result_manager = self.task_manager.task_result_manager
@@ -911,7 +927,6 @@ class TaskServer(
         keeper = self.task_manager.comp_task_keeper
 
         try:
-
             task_id = keeper.get_task_id_for_subtask(subtask_id)
             header = keeper.get_task_header(task_id)
             performance = keeper.active_tasks[task_id].performance
@@ -1007,12 +1022,8 @@ class TaskServer(
         not_accepted = 'not accepted'
 
     def should_accept_provider(  # pylint: disable=too-many-return-statements
-            self,
-            node_id: str,
-            ip_addr: str,
-            task_id: str,
-            provider_perf: float,
-            max_memory_size: int,
+            self, node_id: str, ip_addr: str, task_id: str,
+            provider_perf: float, max_memory_size: int,
             offer_hash: str) -> bool:
 
         node_name_id = short_node_id(node_id)
@@ -1200,10 +1211,7 @@ class TaskServer(
                 del self.forwarded_session_requests[key_id]
                 self.final_conn_failure(data['conn_id'])
 
-    def _get_factory(self):
-        return self.factory(self)
-
-    def _listening_established(self, port, **kwargs):
+    def _listening_established(self, port: int) -> None:
         logger.debug('_listening_established(%r)', port)
         self.cur_port = port
         logger.info(" Port {} opened - listening".format(self.cur_port))
